model/requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
scikit-learn==1.4.2
numpy==2.0.0
tqdm==4.66.4

